+++ buildSrc/src/main/java/com/google/firebase/gradle/plugins/DackkaGenerationTask.kt
@@ -3,20 +3,24 @@ package com.google.firebase.gradle.plugins
 import java.io.File
 import javax.inject.Inject
 import org.gradle.api.DefaultTask
+import org.gradle.api.file.FileCollection
 import org.gradle.api.provider.ListProperty
 import org.gradle.api.provider.Property
 import org.gradle.api.provider.SetProperty
+import org.gradle.api.tasks.CacheableTask
+import org.gradle.api.tasks.Classpath
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.InputFile
 import org.gradle.api.tasks.InputFiles
 import org.gradle.api.tasks.OutputDirectory
+import org.gradle.api.tasks.PathSensitive
+import org.gradle.api.tasks.PathSensitivity
 import org.gradle.api.tasks.TaskAction
 import org.gradle.process.ExecOperations
 import org.gradle.workers.WorkAction
 import org.gradle.workers.WorkParameters
 import org.gradle.workers.WorkerExecutor
 import org.json.JSONObject
-
 /**
  * Extension class for [GenerateDocumentationTask].
  *
@@ -29,25 +33,50 @@ import org.json.JSONObject
  * @property suppressedFiles a list of files to exclude from documentation
  * @property outputDirectory where to store the generated files
  */
+@CacheableTask
 abstract class GenerateDocumentationTaskExtension : DefaultTask() {
-    @get:InputFile
+    @get:[InputFile Classpath]
     abstract val dackkaJarFile: Property<File>
 
-    @get:Input
-    abstract val dependencies: ListProperty<File>
+    @get:[InputFiles Classpath]
+    abstract val dependencies: Property<FileCollection>
 
     @get:InputFiles
+    @get:PathSensitive(PathSensitivity.RELATIVE)
     abstract val kotlinSources: ListProperty<File>
 
     @get:InputFiles
+    @get:PathSensitive(PathSensitivity.RELATIVE)
     abstract val javaSources: ListProperty<File>
 
     @get:InputFiles
+    @get:PathSensitive(PathSensitivity.RELATIVE)
     abstract val suppressedFiles: ListProperty<File>
+
+    @get:InputFiles
+    @get:PathSensitive(PathSensitivity.RELATIVE)
+    abstract val packageListFiles: ListProperty<File>
+
+    @get:Input
+    abstract val clientName: Property<String>
 
     @get:OutputDirectory
     abstract val outputDirectory: Property<File>
 }
+
+/**
+ * Wrapper data class for External package-lists in Dokka
+ *
+ * This class allows us to map package-lists in a type-safe way, versus inline straight to
+ * a map. This extra step could be removed- but it could also catch bugs in the future.
+ *
+ * @property packageList the prepared package-list file to map against
+ * @property externalLink the url to map with when generating the docs
+ */
+data class ExternalDocumentationLink(
+    val packageList: File,
+    val externalLink: String
+)
 
 /**
  * Task to run Dackka on a project.
@@ -64,18 +93,10 @@ abstract class GenerateDocumentationTask @Inject constructor(
     @TaskAction
     fun build() {
         val configFile = saveToJsonFile(constructArguments())
-        launchDackka(configFile, workerExecutor)
+        launchDackka(clientName, configFile, workerExecutor)
     }
 
     private fun constructArguments(): JSONObject {
-        // TODO(b/243675474): Move these to a task input for caching purposes
-        val linksMap = mapOf(
-            "android" to "https://developer.android.com/reference/kotlin/",
-            "google" to "https://developer.android.com/reference/",
-            "firebase" to "https://firebase.google.com/docs/reference/kotlin/",
-            "coroutines" to "https://kotlinlang.org/api/kotlinx.coroutines/kotlinx-coroutines-core/"
-        )
-
         val jsonMap = mapOf(
             "moduleName" to "",
             "outputDir" to outputDirectory.get().path,
@@ -90,9 +111,9 @@ abstract class GenerateDocumentationTask @Inject constructor(
                 "documentedVisibilities" to listOf("PUBLIC", "PROTECTED"),
                 "skipEmptyPackages" to "true",
                 "suppressedFiles" to suppressedFiles.get().map { it.path },
-                "externalDocumentationLinks" to linksMap.map { (name, url) -> mapOf(
-                    "url" to url,
-                    "packageListUrl" to "file://${project.rootDir.absolutePath}/kotlindoc/package-lists/$name/package-list"
+                "externalDocumentationLinks" to createExternalLinks(packageListFiles).map { mapOf(
+                    "url" to it.externalLink,
+                    "packageListUrl" to it.packageList.toURI()
                 ) }
             )),
             "offlineMode" to "true",
@@ -100,6 +121,21 @@ abstract class GenerateDocumentationTask @Inject constructor(
         )
 
         return JSONObject(jsonMap)
+    }
+
+    private fun createExternalLinks(packageLists: ListProperty<File>): List<ExternalDocumentationLink> {
+        val linksMap = mapOf(
+            "android" to "https://developer.android.com/reference/kotlin/",
+            "google" to "https://developers.google.com/android/reference/",
+            "firebase" to "https://firebase.google.com/docs/reference/kotlin/",
+            "coroutines" to "https://kotlinlang.org/api/kotlinx.coroutines/kotlinx-coroutines-core/",
+            "kotlin" to "https://kotlinlang.org/api/latest/jvm/stdlib/"
+        )
+
+        return packageLists.get().map {
+            val externalLink = linksMap[it.parentFile.nameWithoutExtension] ?: throw RuntimeException("Unexpected package-list found: ${it.name}")
+            ExternalDocumentationLink(it, externalLink)
+        }
     }
 
     private fun saveToJsonFile(jsonObject: JSONObject): File {
@@ -111,13 +147,13 @@ abstract class GenerateDocumentationTask @Inject constructor(
         return outputFile
     }
 
-    private fun launchDackka(argsFile: File, workerExecutor: WorkerExecutor) {
+    private fun launchDackka(clientName: Property<String>, argsFile: File, workerExecutor: WorkerExecutor) {
         val workQueue = workerExecutor.noIsolation()
 
         workQueue.submit(DackkaWorkAction::class.java) {
             args.set(listOf(argsFile.path, "-loggingLevel", "WARN"))
             classpath.set(setOf(dackkaJarFile.get()))
-            projectName.set(project.name)
+            projectName.set(clientName)
         }
     }
 }