+++ buildSrc/src/main/java/com/google/firebase/gradle/plugins/ProjectUtils.kt
@@ -14,8 +14,10 @@
 package com.google.firebase.gradle.plugins
 
 import org.gradle.api.Project
+import org.gradle.api.Task
 import org.gradle.api.artifacts.Configuration
 import org.gradle.api.attributes.Attribute
+import org.gradle.api.provider.Provider
 
 fun Project.isAndroid(): Boolean =
         listOf("com.android.application", "com.android.library", "com.android.test")
@@ -41,7 +43,7 @@ val Project.javadocConfig: Configuration
 val Project.dackkaConfig: Configuration
     get() =
         configurations.findByName("dackkaArtifacts") ?: configurations.create("dackkaArtifacts") {
-            dependencies.add(this@dackkaConfig.dependencies.create("com.google.devsite:dackka-fat:1.0.1"))
+            dependencies.add(this@dackkaConfig.dependencies.create("com.google.devsite:dackka-fat:1.0.3"))
         }
 
 /**
@@ -49,7 +51,30 @@ val Project.dackkaConfig: Configuration
  */
 fun Configuration.getJars() = incoming.artifactView {
     attributes {
-        // TODO(b/241795594): replace value with android-class instead of jar after agp upgrade
-        attribute(Attribute.of("artifactType", String::class.java), "jar")
+        attribute(Attribute.of("artifactType", String::class.java), "android-classes")
     }
 }.artifacts.artifactFiles
+
+/**
+ * Utility method to call [Task.mustRunAfter] and [Task.dependsOn] on the specified task
+ */
+fun <T : Task, R : Task> T.dependsOnAndMustRunAfter(otherTask: R) {
+    mustRunAfter(otherTask)
+    dependsOn(otherTask)
+}
+
+/**
+ * Utility method to call [Task.mustRunAfter] and [Task.dependsOn] on the specified task
+ */
+fun <T : Task, R : Task> T.dependsOnAndMustRunAfter(otherTask: Provider<R>) {
+    mustRunAfter(otherTask)
+    dependsOn(otherTask)
+}
+
+/**
+ * Utility method to call [Task.mustRunAfter] and [Task.dependsOn] on the specified task name
+ */
+fun <T : Task> T.dependsOnAndMustRunAfter(otherTask: String) {
+    mustRunAfter(otherTask)
+    dependsOn(otherTask)
+}