+++ buildSrc/src/test/kotlin/com/google/firebase/gradle/plugins/publishing.kt
@@ -29,6 +29,7 @@ data class Project(
     val externalDependencies: Set<Artifact> = setOf(),
     val releaseWith: Project? = null,
     val customizePom: String? = null,
+    val publishJavadoc: Boolean = false,
     val libraryType: LibraryType = LibraryType.ANDROID
 ) {
     fun generateBuildFile(): String {
@@ -42,6 +43,7 @@ data class Project(
             firebaseLibrary {
                 ${if (releaseWith != null) "releaseWith project(':${releaseWith.name}')" else ""}
                 ${if (customizePom != null) "customizePom {$customizePom}" else ""}
+                ${"publishJavadoc = $publishJavadoc"}
             }
             ${if (libraryType == LibraryType.ANDROID) "android.compileSdkVersion = 26" else ""}